Rename angular.component files and update build configurations
