+++ angular/demo/set-number/set-number.component.ts
@@ -1,10 +1,10 @@
 import { Component, ViewChild } from '@angular/core';
-import { IntlTelInputComponent, PHONE_ERROR_MESSAGES } from '../../src/intl-tel-input/angularWithUtils.component';
+import IntlTelInputComponent, { PHONE_ERROR_MESSAGES } from '../../src/intl-tel-input/angularWithUtils';
 
 @Component({
   selector: "app-root",
   template: `
-    <form>
+    <div>
       <intl-tel-input
         #telInput
         (numberChange)="handleNumberChange($event)"
@@ -23,7 +23,7 @@ import { IntlTelInputComponent, PHONE_ERROR_MESSAGES } from '../../src/intl-tel-
       @if (notice) {
         <div class="notice">{{ notice }}</div>
       }
-    </form>
+    </div>
   `,
   standalone: true,
   imports: [IntlTelInputComponent]
