+++ angular/demo/toggle-disabled/toggle-disabled.component.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { IntlTelInputComponent } from '../../src/intl-tel-input/angularWithUtils.component';
+import IntlTelInputComponent from '../../src/intl-tel-input/angularWithUtils';
 
 @Component({
   selector: "app-root",
